+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## [Unreleased]
+
+## Added
+
+- Added `alwaysVisible` property to **Scrollable**.
+
 ## [1.16.1] - 2025-05-22
 
 ### Fixed